+++ lib/funcs/ping.js
@@ -1,0 +1,5 @@
+import { Discord, On, Client } from '@typeit/discord';
+import { Message, Attachment, TextChannel } from 'discord.js';
+module.exports = function(message){
+	message.answer("PONGGERS!")
+}
